Fix extensions and move them to their package
